requirements.txt
# Python dependencies for the observable AI chatbot
# Versions are pinned for reproducible builds

# FastAPI web framework
fastapi==0.104.1

# ASGI server for running FastAPI applications
uvicorn[standard]==0.24.0

# OpenAI SDK for communicating with OpenRouter API
openai==1.3.7

# Prometheus client library for Python metrics
prometheus-client==0.19.0

# Python-dotenv for loading environment variables from .env files
python-dotenv==1.0.0

# HTTP client library (used by FastAPI and OpenAI SDK)
httpx==0.25.2

# Pydantic for data validation (FastAPI dependency)
pydantic==2.5.0
